First commit of TiddlyFox 2
A new version of TiddlyFox, now renamed to “TiddlyWiki for Firefox”.

It is a complete rewrite, now based on the latest version of Mozilla’s
Add-on SDK.

Initial features include a toolbar button and dropdown panel that are
used to explicitly enable saving, instead of the old alert system. It’s
also possible to later revoke permission for a file.

I intend to clean up the user interface and add support for backups
before the first release. Other features including clipping will follow.