Reverting to 1.0alpha20
Sadly, all the work of 8763a0c232beff54d61c0a823c46400e28ed5408 has to
be reverted. The trouble is that the latest version of Mozilla’s
Add-on SDK is no longer compatible with Firefox for Android (thanks to
@buggyj for the discovery).

Instead, we’ll pick up the older version, add support for backups and
work towards a 3.0.0 release that works with the latest Firefox
extension process architecture.

I’m also hoping to explore alternate configurations, such as perhaps
running (parts of) the Node.js configuration within Firefox.